add a little bit logging remove a little bit of csv
sqlite version seems much slower on NAS, wonder if there's a good way to speed it up? 

wild theories about what is slowing it down:
- I'm misremembering how fast the other version was because I didn't write it down
- inserting one row at a time in a loop isn't fast
- sqlite is that much slower than the filesystem

that last point though is maybe refuted by this, though: https://www.sqlite.org/fasterthanfs.html

for the first point, I should check out the last point in history that had that version and actually write it down.

for the middle point, look how annoying it would be to batch things.