play: add volume control
also add an `install` script to toss some bins in my home bin dir.

ergonomics-wise, I'd like to be able to control the volume without restarting the application, particularly since I had to reset back before I added the ability to skip a # of samples when starting.

I can probably do this by: 
- setting up an input reader thread
- turning Volume into an AtomicF32 from the atomic_float package.
- storing a new Volume when receiving input
- reading Volume on each audio frame

for extra credit, I could interpolate the change over the samples of the frame. so, for example, going from 0.1 to 1.0 would happen over the course of the frame rather than instantly.

this may not be necessary: since frames are 1/20th of a second right now, the change might sound instantaneous either way.