play: rudimentary live volume control
it more or less works! takes a number 0-100 on stdin and sets that into an atomic that gets loaded in the audio render thread.

I should figure out how to benchmark stuff so I can measure the difference between using an atomic here vs using a locked struct that can hold more state.