can skip to a certain point, but it's clunky
the way it's implemented now is that it still reads every single packet on the way to the location.

we should have the reader take a tracklist instead of a vec of paths, that way it can be smarter about skipping.

this is going to require having something like a TrackListEntry which includes a Track and the start and end points of the track.