increase sample channel buffer
for tracks that are 96kHz we need a larger buffer when reading off NAS

I should also look at better ways of filling the buffer in chunks. I'm using a sleep now when the buffer is full, rather than blocking until it's ready, because filling it in chunks seemed more CPU efficient.

I'm not sure I got that right, though! I think when I looked at it I was running in debug mode, not release.